removed circular dependencies between ember-metal/properties and ember-metal/property_set
