+++ packages/ember-views/lib/views/view.js
@@ -24,7 +24,7 @@ import {
 } from "ember-metal/utils";
 import { isNone } from 'ember-metal/is_none';
 import { Mixin } from 'ember-metal/mixin';
-import { deprecateProperty } from "ember-metal/properties";
+import { deprecateProperty } from "ember-metal/deprecate_property";
 import { A as emberA } from "ember-runtime/system/native_array";
 
 import { dasherize } from "ember-runtime/system/string";
